+++ firmware/ga23_dev_bmp085.h
@@ -5,7 +5,7 @@
 #define GA23_DEV_BMP085_H
 void ga23_dev_bmp085_open(void);
 int ga23_dev_bmp085_avail(void);
-int ga23_dev_bmp085_read(void);
-float ga23_dev_bmp085_read_temp(void);
+uint32_t ga23_dev_bmp085_read(void);
+uint16_t ga23_dev_bmp085_read_temp(void);
 #endif
